Fix stream-chat client mocking
Some recent changes on stream-chat introduced a new way to handle users
and tokens. This led to some of our tests breaking, as our mocking
implementation no longer complied with the actual API of the library.

Here we fix this by mocking and abstracting away the bits that deal with
user and tokens.

Signed-off-by: Vini Andrade <vini.andrade.dev@gmail.com>