Split shouldMessageComponent update util function
This splits the shouldMessageComponent update into two methods, one for
comparing the props and telling if they're equal or not and the other
with the responsibility of telling the component to update or not based
on the first.

This will make it less confusing to use this utility function when
memoizing functional components, as the callback for `React.memo` needs
to know if previous and next props are equal or not, rather then telling
if the component should update.