Improve edit handler
Adding the whole editing state handling to the edit handler hook, as
this now belongs to the message components. While doing it, I noticed
that there are 2 edit handlers passed to the MessageActions component,
both intended to deal with setting the edit state to true. Because of
that, and because for now this component is only used internally, I
removed one of them in favor of the other, simplifying the logic there.