Sanitize URL image sources in Image component
Some outdated browsers could potentially still allow javascript execution from img
elements, which could lead to XSS. Although such browsers would fall
outside the scope of our supported browsers, fixing the vulnerability is
a simple, so we're opting for doing it here.

It is important to note that the javascript check will be obsolete in
the next major react version, where it will no longer render uris with
javascript as a protocol.