Fix an error when a hyphen causes a false positive
The `search` function on `NimbleEmojiIndex` can return an array with
a null value. This in turn causes an error when attempting to
destructure that element later on.

I wasn't able to write a test that highlights this error. I found it
difficult to replace the dependency on NimbleEmojiIndex with a mock.

Open to pointers on how to do that so I can complete this with an actual
failing test.

In the meantime, ensuring that `found` only contains non-null values
will fix the error.