fix: detect mentions of users who have email as their name
The link detection process kicks-in before user mentions are detected.
With this change, while detecting links we also take care whether the detected link
matches to an email which is identical to possibly mentioned user's name.

As part of this change, mdast-util-find-and-replace is updated to it's latest version
as the one we were previously using had a bug, and sometimes it could skip matching nodes.