fix: include mdast-util-find-and-replace into our CJS bundle (#1702) (#1703)
Our recent upgrade of `mdast-util-find-and-replace` in #1698 broke our CommonJS bundle.
The `mdast-util-find-and-replace` package is now bundled within our CJS bundle. This will eliminate the conflict between ESM and CJS modules in NodeJS.
As part of this change, a very naive smoke test is introduced aiming for earlier detection of such incompatibilities.