fix: do not try close message input's autocomplete dropdown on textarea blur (#2015)
The blur event's relatedTarget value in Safari is incorrect when clicking on autocomplete item. We cannot therefore decide, whether to close or not the autocomplete dropdown. This fix introduces change that will lead to ignoring the textarea blur event in Safari. That means that user will be able to close the autocomplete dropdown only by selecting the item.