fix: make cooldown timer to rely on skip-slow-mode capability  (#2018)
Adapt the cooldown timer enablement to the latest bug-fix on the backend:

The 'slow-mode' own capability was replaced by 'skip-slow-mode'. This capability should only be used to determine if the user can skip cooldown mode and not to determine if the channel has cooldown enabled. The channel.own_capabilities array will contain capability 'skip-slow-mode' if a user is assigned a role that has skip-channel-cooldown permission.

Cooldown mode enablement is signaled by the presence of channel.cooldown value (num of seconds). When cooldown mode is enabled a user can send a message only if:

- in the cooldown period if channel.own_capabilities includes 'skip-slow-mode'
- the time elapsed since the last message sent is > channel.cooldown
- no previous message was sent by the given user in the given channel

Otherwise the user has to wait until the cooldown period has elapsed before sending next message.