chore: Git Hooks & general cleanup (#2192)
### 🎯 Goal

Personally - not a big fan of Git Hooks - but I believe the general
consensus is to keep them so I did some upgrading/cleaning up. I also
removed the Python script as we're not commiting any binaries to our SDK
and I'm not installing Python on my machine to run it.

Also - I'm considering ditching the `commitlint` extension as all of our
PR's are squashed and that commit message is being checked by the CI
(Check PR title).

Removed some of the other files that I believe are unused and no longer
relevant.