feat: show full list of reactions in a modal (#2249)
### 🎯 Goal

We're updating UI for displaying reactions. Instead of showing the
latest reactions only, we're adding a separate modal window full the
full list of reactions.

Although not a breaking change in terms of API, it's a breaking change
in terms of UI and styling.

### 🛠 Implementation details

1. New component `ReactionsListModal` is added (rendered by default
`ReactionsList`).
2. New handler added on message context level: `handleFetchReactions`.
It's expected to load all reactions for current message (e.g. using our
paged endpoint for fetching reactions). As usual, it can be overriden on
`Message` or `ReactionsList` level.
3. Instead of showing counts for the latest reactions only, we show
counts for all supported reactions in the `ReactionsList`.

### 🎨 UI Changes


![image](https://github.com/GetStream/stream-chat-react/assets/975978/69b93a0b-571f-4823-9f21-ceb744cfb4af)

### To-Do

- [x] Tests for ReactionsListModal
- [x] Add translations