feat: use lint-staged for lint and fixes (#2258)
### 🎯 Goal

I got a bit frustrated with our pre-commit hook failing because changes
in files unrelated to the commit. Plus, the command ran by the
pre-commit hook was pretty slow, running for a couple of seconds.

It would be nice for the pre-commit hook to only lint the files that are
actually staged.

### 🛠 Implementation details

The easiest (and most standard) way to achieve that is by using the
`lint-staged` package. It allows to pass a list of **staged** files
(filtered by glob) to our lint commands.

That prevents the pre-commit hook from failing because of errors in
files that are not part of the commit. And, since we don't lint the
whole codebase every time, it runs faster too.

I also added a new command `fix-staged` that will run auto-fixes for the
staged files only. The pre-commit hook now prompts to use this command.

### Demo


https://github.com/GetStream/stream-chat-react/assets/975978/f36a7109-92e7-45a5-b431-1e99a8538285