feat: implement message bounce flow (#2254)
### 🎯 Goal

🚂 https://github.com/GetStream/stream-chat-js/pull/1213
🚂 https://github.com/GetStream/stream-chat-css/pull/264

In chats with moderation rules set up, message can bounce if its content
is deemed potentially harmful. The author of a bounced message should
then be presented with four alternatives:

1. Edit the message and try sending it again
2. Try sending it again as-is (this is helpful for "bounce then flag"
flow)
3. Remove the message
4. Do nothing. Bounced messages are ephemeral, so it will soon disappear
on its own

### 🛠 Implementation details

This PR introduces a couple of new components, including the
`MessageBounceModal` which is rendered by `MessageSimple` when a bounced
message is clicked.

The contents of the modal (`MessageBounceOption`) is an overridable
component that should ideally render three buttons for the first three
alternative options listed above. The callbacks for said buttons are
provided via `MessageBounceContext`.

### 🎨 UI Changes

The chat in the screenshot has a semantic filter set up which is
triggered by the word "midnight". Here's what a bounced message with the
word "midnight" looks like:


![image](https://github.com/GetStream/stream-chat-react/assets/975978/9476cde6-f310-41a8-bb6d-3e7ed0f58421)

Clicking on the bounced messages opens `MessageBounceModal`:


![image](https://github.com/GetStream/stream-chat-react/assets/975978/9bbb7201-3342-4ce7-88f2-9a3131c96878)

Clicking "Edit Message" opens the standard editing UI:


![image](https://github.com/GetStream/stream-chat-react/assets/975978/ef627801-9b3b-46e5-8d12-eb8b3fd08afd)

### To-Do and Next Steps

[The design
doc](https://www.figma.com/file/ekifwChR9tR7zRJg1QEzSM/Chat-UI-Kit-1.0-All-platforms?type=design&node-id=23638-313355&mode=design)
for this feature also features a notification banner with a button,
which is displayed when a message bounces. Clicking the button should
bring the user to the bounced message.

We don't have a way to have interactive elements within channel
notifications at the moment, but this is going to be implemented in
further PRs in two steps:

1. Allow passing arbitrary JSX to the notification, not just text
2. Implement a bounced message notification with a button to bring the
user to the message

- [x] Release `stream-chat-css` with udpated styles
- [x] Release `stream-chat-js` with updates types
- [x] Cover `MessageBounceModal` and `MessageBounceOptions` with tests
- [x] Document new components and customization options
- [x] Document the moderation flow