feat: add customizable reactions sorting (#2289)
### 🎯 Goal

Currently the order of reactions in the `ReactionsList` depends of the
order of keys in the `reaction_counts` object returned by the backend.
This leads to unstable reaction order, which can change on page reloads,
or when the list of reactions changes.

We want to provide a way to customize this behavior, and to have a
stable order of reactions by default.

### 🛠 Implementation details

Added a new `MessageContext` property: `sortReactions`. It accepts two
reaction objects, and should return a positive/negative/zero numeric
value, similar to the `Array.prototype.sort` method.

The `sortReactions` prop can be passed both to `MessageList` and
`VirtualizedMessageList`, as well as to the `Message` component
directly. It will then be available via the `MessageContext`.