fix: mark retryable duplicated messages as received (#2331)
### 🎯 Goal

In some circumstances, the response about the delivery of a message may
be aborted (timeout) or the web socket information about `message.new`
might not arrive because of network conditions (offline, etc...).
In this case, the SDK will prompt the user to re-send the message. Once
the user retries, the backend will error with `SendMessage failed:
Message with id XYZ already exists`.

This error wasn't handled properly by the SDK and it led the user to
re-attempt sending the same message over and over again in a loop.
This loop could have been broken only by reloading the app. This PR aims
to fix that.

### 🛠 Implementation details

The SDK now recognizes this error, and once it happens, it immediately
marks the message as `received`. This breaks the re-send loop for the
user. Any UI inconsistencies (ordering of messages) related to the
network glitches and optimistic rendering will be sorted out once the
network connection restores and our SDK establishes a solid connection
to our backend systems.