fix: suggestion onClickHandler type (#2343)
Noticed that an example in our docs is broken. Turns out, we provided
wrong typings for the `onClickHandler` in `ChatAutoComplete`.

The reason it passed type checks before is that the suggestion list
component is in a JSX file (not TSX). I'm not going to switch it TSX
right now, but the types are worth fixing quickly.

Also removed outdated demos that will break because of this change.