feat: adopt new queryReactions endpoint (#2388)
🚂 https://github.com/GetStream/stream-chat-js/pull/1279

The new `queryReactions` endpoint supports backend filtering and
sorting, so now we don't have to fetch all reactions and then process
them on the client.

`sortReactionDetails` prop for client-side filtering is still supported,
but deprecated. The new recommended way is to pass `reactionDetailsSort`
object which will be passed in the reactions query.