fix!: extract MP3 encoder plugin (#2447)
BREAKING CHANGE: MP3 audio encoder has to be explicitly imported and
used as a plugin for audio recordings. The default audio recording
format is audio/wav.
BREAKING CHANGE: @breezystack/lamejs became a peer dependency and has to
be installed by the integrator so that the MP3 audio encoder can work
properly.