fix: quote replies in threads (#2487)
### 🎯 Goal

It's perfectly normal to quote a message that is part of a thread, as
long as it's quoted in the same thread. Looks like this logic has been
broken for some time. This PR fixes it.

### 🛠 Implementation details

Take a look at how quoted message is handled in `Channel`:

```
const messageData = {
  // ...
  quoted_message_id: parent_id === quotedMessage?.parent_id ? quotedMessage?.id : undefined
};
```

Quoting is allowed as long as the quoted message is part of the same
thread, or both are regular non-thread messages.

However, there was a mismatch in how `MessageInputFlat` displays a
quoted message: it was always hidden for thread messages, which lead to
confusing behavior.

Now the logic is the same in both places.