fix: reset message composer state partially when poll message is sent (#2703)
When sending a poll message (contains only poll id) the composer id has
to be created anew (for the next msg) and the poll composer state has to
be reset. The rest of the states should be kept as they can be used for
the next message.

This should be moved to the LLC in the future, once
`MessageComposer.sendMessage()` exists.