feat: support overriding components in MessageList (#2890)
### 🎯 Goal

This PR aligns overriding functionality between the
VirtualizedMessageList (VML) and the regular MessageList (ML).

The goal is to be able to integrate components like React Aria's
[GridList](https://react-spectrum.adobe.com/react-aria/GridList.html)
into ML.

Currently, VML supports overriding the list and item components via
`additionalVirtuosoProps.components`. This PR adds the same ability to
ML via the ComponentContext.

### 🛠 Implementation details

We add two more components to ComponentContext:
- MessageListWrapper wraps all message list items (default is `ul`).
- MessageListItem wraps each message list item (default is `li`).

These components are consumed in MessageList and in `renderMessages()`
respectively.

This makes the following usage possible:

```jsx
<Chat client={client}>
  <Channel channel={channel}>
    <ComponentProvider value={{
      MessageListItem: (props) => (
        <div data-testid='message-list-item' {...props} />
      ),
      MessageListWrapper: (props) => (
        <div data-testid='message-list-wrapper' {...props} />
      ),
    }}>
      <MessageList {...msgListProps} />
    </ComponentProvider>
  </Channel>
</Chat>
```

Sometimes it's also valuable to be able to access the current
`RenderedMessage` item from MessageListItem component. In VML it's
possible by accessing Virtuoso's context to get `processedMessages` and
using the `data-index` attribute to get item's index.

To align this functionality with ML, this PR adds `processedMessages` to
MessageListContext, and adds `data-index` attribute to message list
items.

### 🎨 UI Changes

None by default.